edit init-server script

+++ init-server.sh
@@ -2,18 +2,25 @@
 
 echo "-----------------Start init this server----------------"
 git clone https://github.com/CrazyZPL/nvim-config.git
-mv nvim-config .config
+mv nvim-config ~/ && mv nvim-config .config
 
-echo "Start setup node"
-curl -sL https://deb.nodesource.com/setup_16.x | sudo -E bash -
-apt-get install nodejs -y
-echo "---------------------End setup node---------------------"
+if node -v; then
+  echo "node is found"
+else
+  echo "Start setup node"
+  curl -sL https://deb.nodesource.com/setup_16.x | sudo -E bash -
+  apt-get install nodejs -y
+  echo "---------------------End setup node---------------------"
+fi
 
-echo "----------------------Start setup neovim-----------------"
-sudo add-apt-repository ppa:neovim-ppa/stable -y
-sudo apt-get update -y
-sudo apt-get install neovim -y
-echo "------------------------End setup neovim-----------------------"
+if nvim; then
+  echo "neovim is found"
+else
+  echo "----------------------Start setup neovim-----------------"
+  sudo add-apt-repository ppa:neovim-ppa/stable -y
+  sudo apt-get update -y
+  sudo apt-get install neovim -y
+  echo "------------------------End setup neovim-----------------------"
 
 echo "----------------------Start setup Packer-----------------------"
 git clone --depth 1 https://github.com/wbthomason/packer.nvim \
@@ -32,8 +39,11 @@ apt install gcc g++ -y
 apt install unzip -y
 echo "------------------end setup gcc&g++------------------------"
 
-echo "------------------Start setup Golang------------------------"
-wget https://go.dev/dl/go1.19.1.linux-amd64.tar.gz
-rm -rf /usr/local/go && tar -C /usr/local -xzf go1.19.1.linux-amd64.tar.gz
-echo "export PATH=/usr/local/go/bin:$PATH" >> ~/.bashrc
-echo "--------------------End setup Golang------------------------"
+if go version; then
+  echo "Golang is found"
+else
+  echo "------------------Start setup Golang------------------------"
+  wget https://go.dev/dl/go1.19.1.linux-amd64.tar.gz
+  rm -rf /usr/local/go && tar -C /usr/local -xzf go1.19.1.linux-amd64.tar.gz
+  echo "export PATH=/usr/local/go/bin:$PATH" >> ~/.bashrc
+  echo "--------------------End setup Golang------------------------"
